requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
bidict==0.23.1
build==1.2.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dill==0.3.8
distro==1.9.0
docutils==0.21.2
fastapi==0.110.3
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
importlib_metadata==8.4.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.7.1
keyring==25.3.0
lazy_loader==0.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.4.0
nh3==0.2.18
openai==1.54.4
packaging==24.1
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.2.2
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
readme_renderer==44.0
redis==5.0.8
reflex==0.6.4
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.13
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.3
setuptools==75.4.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlmodel==0.0.21
starlette==0.37.2
starlette-admin==0.14.1
tabulate==0.9.0
tomlkit==0.13.2
tqdm==4.67.0
twine==5.1.1
typer==0.12.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.23.0
websockets==13.0
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
zipp==3.20.0
